+++ src/popups/ReportLevel.hpp
@@ -8,7 +8,7 @@
 
 using namespace geode::prelude;
 
-class ReportLevel : public Popup<> {
+class ReportLevel : public Popup<>, public FLAlertLayerProtocol {
     GJGameLevel* m_level{ nullptr };
     TextInput* m_input{ nullptr };
     CCMenuItemSpriteExtra* sendButton{ nullptr };
@@ -19,8 +19,17 @@ class ReportLevel : public Popup<> {
 
         auto size = m_mainLayer->getContentSize();
 
+        auto infoSprite = CCSprite::createWithSpriteFrameName("GJ_infoIcon_001.png");
+        auto infoBtn = CCMenuItemSpriteExtra::create(
+            infoSprite, this, menu_selector(ReportLevel::onInfo)
+        );
+        infoBtn->setPosition({ size.width - 20.f, size.height - 20.f });
+        auto infoMenu = CCMenu::createWithItem(infoBtn);
+        infoMenu->setPosition({0, 0});
+        m_mainLayer->addChild(infoMenu);
+
         auto info = CCLabelBMFont::create(
-            "Only report for serious issues such as inappropriate content, harassment, etc.\nDO NOT beg for rates here, we aren't associated with RobTop.",
+            "Only report for serious issues and use the correct report type.\nPress the INFO button for more info.",
             "chatFont.fnt"
         );
         info->setColor({255, 255, 255});
@@ -28,13 +37,22 @@ class ReportLevel : public Popup<> {
         info->setPosition({ size.width / 2, size.height / 2 + 55.f });
         m_mainLayer->addChild(info);
 
-        auto warning = CCLabelBMFont::create("Repeated misuse of AbuseDB will lead to a permanant ban.", "chatFont.fnt");
+        auto warning = CCLabelBMFont::create("Repeated misuse of AbuseDB will lead to a permanent ban.", "chatFont.fnt");
         warning->setColor({0, 255, 255});
         warning->setScale(0.9f);
         warning->setPosition({ size.width / 2, 30.f });
         m_mainLayer->addChild(warning);
 
-        m_input = TextInput::create(260.f, "Enter reason for report. You should also provide your username here\nin case we need to contact you. (turn messages on)", "chatFont.fnt");
+        auto winSize = CCDirector::sharedDirector()->getWinSize();
+        auto discordLabel = CCLabelBMFont::create("Join our Discord Server at https://dsc.gg/devlin for updates or for questions about AbuseDB.", "chatFont.fnt");
+        discordLabel->setColor({255, 255, 255});
+        discordLabel->setScale(0.8f);
+        discordLabel->setAnchorPoint({0.5f, 0.0f});
+        discordLabel->setPosition({ winSize.width / 2, 40.f });
+        discordLabel->setOpacity(220);
+        this->addChild(discordLabel, 1000);
+
+        m_input = TextInput::create(260.f, "Enter a reason for your report. Please add detail so we don't decline your report.", "chatFont.fnt");
         m_input->setPosition({ size.width / 2, size.height / 2 + 20.f });
         m_mainLayer->addChild(m_input);
 
@@ -49,6 +67,38 @@ class ReportLevel : public Popup<> {
         return true;
     }
 
+    void onInfo(CCObject*) {
+        auto winSize = CCDirector::sharedDirector()->getWinSize();
+        FLAlertLayer::create(
+            this,
+            "Information",
+            "<cg>page 1</c>",
+            "Next Page",
+            nullptr,
+            winSize.width - 40.f,
+            true,
+            winSize.height - 20.f,
+            1.0f
+        )->show();
+    }
+
+    void FLAlert_Clicked(FLAlertLayer*, bool btn2) override {
+        if (!btn2) {
+            auto winSize = CCDirector::sharedDirector()->getWinSize();
+            FLAlertLayer::create(
+                nullptr,
+                "DOs and DON'Ts",
+                "<cg>page 2</c>",
+                "OK",
+                nullptr,
+                winSize.width - 40.f,
+                true,
+                winSize.height - 20.f,
+                1.0f
+            )->show();
+        }
+    }
+
     void onSend(CCObject*) {
         if (!m_level) { return; }
         if (!GJAccountManager::get()) { return; }
@@ -57,11 +107,18 @@ class ReportLevel : public Popup<> {
 
         auto reason = m_input->getString();
 
+        std::string reporter = "unknown";
+        int reporterID = 0;
+        if (GJAccountManager::get()) {
+            reporter = GJAccountManager::get()->m_username;
+            reporterID = GJAccountManager::get()->m_accountID;
+        }
+
         web::WebRequest req;
         req.header("Content-Type", "application/x-www-form-urlencoded")
-           .bodyString(fmt::format("level_id={}&level_reason={}",
+           .bodyString(fmt::format("level_id={}&level_reason={}&my_username={}&my_playerID={}",
                                    m_level->m_levelID.value(),
-                                   reason));
+                                   reason, reporter, reporterID));
         m_reportListener.bind([this](web::WebTask::Event* e) {
             if (e->getProgress())
                 return;
@@ -69,15 +126,15 @@ class ReportLevel : public Popup<> {
                 return;
             if (auto res = e->getValue()) {
                 if (res->code() == 403) {
-                    FLAlertLayer::create("Banned", "You have been banned from reporting.\nIf you have any real reports then consider joining our Discord Server at dsc.gg/devlin.", "OK")->show();
+                    FLAlertLayer::create("Banned", "You have been banned from reporting.\nYou can appeal by joining our Discord Server at dsc.gg/devlin and begging on your knees.", "OK")->show();
                     this->onClose(nullptr);
                     return;
                 }
                 if (res->code() >= 200 && res->code() < 300) {
-                    FLAlertLayer::create("Success", "Report sent successfully, we are taking a look at it.", "OK")->show();
+                    FLAlertLayer::create("Success", "Report sent successfully, we are taking a look at it now. Just in case we need more information, please have your messages on.", "OK")->show();
                     this->onClose(nullptr);
                 } else {
-                    FLAlertLayer::create("Error", "Failed to send report. Please try again.", "OK")->show();
+                    FLAlertLayer::create("Error", "Failed to send report. Please try again laters.", "OK")->show();
                     if (sendButton) sendButton->setEnabled(true);
                 }
             }