+++ changelog.md
@@ -1,3 +1,12 @@
+# v1.2.0
+- abusedb menu (replaces the View Flags button on MenuLayer)
+- usernames marked as red in comments if they are flagged on AbuseDB
+- to support the previous change mentioned above, we have added usernames to the AbuseDB flag list and the mod fetches a full list of usernames on launch
+- added more information to Report UI's
+- added discord server link to Report UI's just in case for people who want to ask
+- made abusedb menu button smaller in size
+- the reporters username and AccosuntID are forced to be sent when reporting, really sick of people nowadays
+- abusedb mods have text color now
 # v1.1.1
 - ban system
 - better labels on report ui's since people do NOT know what they are doing at all